requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
scikit-optimize>=0.10
PyYAML>=6.0
torch>=2.1
tqdm>=4.66
